requirements.txt
streamlit
pandas
numpy
plotly
requests
yfinance
yahoo_fin
python-dotenv
finnhub-python